Refactor classes to support multiple underlays.
