+++ api/src/test/java/bio/terra/tanagra/api/omop/OmopBrandTest.java
@@ -1,0 +1,55 @@
+package bio.terra.tanagra.api.omop;
+
+import bio.terra.tanagra.api.BaseQueryTest;
+import bio.terra.tanagra.api.EntityQueryRequest;
+import bio.terra.tanagra.api.entityfilter.TextFilter;
+import bio.terra.tanagra.query.filtervariable.FunctionFilterVariable;
+import bio.terra.tanagra.testing.GeneratedSqlUtils;
+import bio.terra.tanagra.underlay.Underlay;
+import java.io.IOException;
+import org.junit.jupiter.api.Test;
+
+public abstract class OmopBrandTest extends BaseQueryTest {
+  @Test
+  void noFilter() throws IOException {
+    EntityQueryRequest entityQueryRequest =
+        new EntityQueryRequest.Builder()
+            .entity(getEntity())
+            .mappingType(Underlay.MappingType.INDEX)
+            .selectAttributes(getEntity().getAttributes())
+            .limit(DEFAULT_LIMIT)
+            .build();
+    GeneratedSqlUtils.checkMatchesOrOverwriteGoldenFile(
+        querysService.buildInstancesQuery(entityQueryRequest).getSql(),
+        "sql/" + getSqlDirectoryName() + "/brand-noFilter.sql");
+  }
+
+  @Test
+  void textFilter() throws IOException {
+    // filter for "brand" entity instances that match the search term "paracetamol"
+    // i.e. brands that have a name or synonym that includes "paracetamol"
+    TextFilter textFilter =
+        new TextFilter.Builder()
+            .textSearch(getEntity().getTextSearch())
+            .functionTemplate(FunctionFilterVariable.FunctionTemplate.TEXT_EXACT_MATCH)
+            .text("paracetamol")
+            .build();
+
+    EntityQueryRequest entityQueryRequest =
+        new EntityQueryRequest.Builder()
+            .entity(getEntity())
+            .mappingType(Underlay.MappingType.INDEX)
+            .selectAttributes(getEntity().getAttributes())
+            .filter(textFilter)
+            .limit(DEFAULT_LIMIT)
+            .build();
+    GeneratedSqlUtils.checkMatchesOrOverwriteGoldenFile(
+        querysService.buildInstancesQuery(entityQueryRequest).getSql(),
+        "sql/" + getSqlDirectoryName() + "/brand-textFilter.sql");
+  }
+
+  @Override
+  protected String getEntityName() {
+    return "brand";
+  }
+}
